getting-started: provide path for running without Ignition config
To lower the barrier a bit more on trying FCOS, tweak the AWS docs to
clarify that one doesn't *have* to specify an Ignition config. Provide
an example using `--key-name`.

There's also the live image... we should document that too at some
point.

This came out of feedback from a Reddit comment:
https://www.reddit.com/r/Fedora/comments/epyzbx/fedora_coreos_out_of_preview/feoysmk