sysconfig-network-configuration: remove line wrap
Wrapping lines of "NOTE:" blocks results in translation strings with
leading whitespace chars and trailing newline chars. It makes it harder
to translate the documentation because the translator has to deal with
adding or not the newlines, and where (if add); Also, Weblate notifies
if number of lines don't match, so it is quite annoying.

Other FCOS docs doesn't wrap any lines, so this commit unwraps all
lines in this document.