Mention "OS extensions" in package layering docs
I like the "OS extensions" framing of looking at package layering to
make it more explicit what the intent should be. Let's add that in the
docs to help users discern when it's appropriate to use layering (though
I agree it's not always clear and some people just really want their vim
directly on the host and that's fine).