tutorial-services: Ensure issue file written before agetty
We need to write the issue file into `/etc/issue.d` before
`systemd-user-sessions.service` to ensure that it is written before
`agetty` is started so that `agetty` will display our issue snippet.
We can either do this or have `agetty --reload` in the script, but
this is cleaner IMO and it matches console-login-helper-messages.
https://github.com/coreos/console-login-helper-messages/pull/100