provisioning/exoscale: overhaul exoscale documentation
The exoscale CLI had gone through quite a few revisions and old
commands no longer worked. Also the FCOS images that are in Exoscale
are Fedora 34 based. Let's stop referring to the images in Exoscale
and just rely on the Custom Template creation process for our users
for now until we can start publishing the images ourselves.