*: s/serial console/console/ where the serial part doesn't matter
We reference the serial console in a number of places.  Sometimes it
refers to the "serial" part specifically (e.g. how to exit the QEMU serial
console, or if Butane configs specifically enable autologin there), but
in a few places the "serial" part is an implementation detail.  In the
latter case, just call it the console.