ibmcloud: update ibmcloud cli provisioning steps
Changes addressed in this commit:
- The ibmcloud container link was broken and I could not find the
doc anymore so the instructions have been updated to the manual
installatin of ibmcloud CLI.

- A resource group is required for creating all the resources
necessary to provision a FCOS system.

- I got tripped up on the @ symbol for passing the ignition config.
I thought the whole thing was supposed to be replaced with the path
to the ignition config but the @ is required before the path. Make
it obvious.