tutorials: use ssh_authorized_keys_local
It's a bit more ergonomic for the user when copy/pasting to not have
to then edit the contents of the file and paste in their public key
each time. Since we have `ssh_authorized_keys_local` now let's just
use that and have the tutorial setup steps instruct the users to create
a local `ssh-key.pub` before they start.